[FEAT]: ChangePasswordCoordinator 로 명칭 변경 및 구현
